requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pymc>=5.0.0
python-dotenv>=1.0.0
slowapi>=0.1.9

